Modify the commit workflow to allow some unrecorded reviews
This is connected to the effort to have releases being staged automatically.
The automaton that performs this sort of staging cannot know who's going to
be the reviewer, so our previous habit of pre-setting the commits with the
intended reviewer is not possible to do any more, and changing them later
with addrev would remove any signature the automaton adds, for example on
the release tag.

We therefore need to make an exception with such commits.

Fixes #49