Update platform information related to AIX
Related to https://github.com/openssl/openssl/pull/24609 PR,
a new configuration is added for aix platform to support openxl compiler.
This includes the new platform in the platform-list link.

The aix platforms are moved from unadopted to community platforms with
@sanumesh as volunteer.

The aix64-cc-solib and aix-cc-solib platforms are kept as unadopted
platforms as they seem to be related to customer specific library packaging.